Python 3 and more
* Port to Python 3
* Use `fInitializeProduct.py`
* Use `mExports.py`
* dxProductDetails version 2021-07-02 11:16
* Standardize `Tests.cmd`
* Improve `Tests\fTestDependencies.py`
* Standardize variable/function/method naming convention

* Improve ASan support
* Standardize code for various things
* Handle user pressing CTRL+C
* Improve stack processing code to better deal with incorrect frames.
* Improve bug translation code and translations data
* Improve test assembly code
* Add Unicode DLL for testing

+++ ftuLimitedAndAlignedMemoryDumpStartAddressAndSize.py
@@ -33,7 +33,7 @@ def ftuLimitedAndAlignedMemoryDumpStartAddressAndSize(uMostInterestingAddress, u
     else:
       # The size before and after the address are both too large: the memory dump will have the most interesting
       # address in the middle.
-      uMemoryDumpStartAddress = long(round(uMostInterestingAddress - dxConfig["uMaxMemoryDumpSize"] / 2));
+      uMemoryDumpStartAddress = int(round(uMostInterestingAddress - dxConfig["uMaxMemoryDumpSize"] / 2));
       # Align new start address:
       uMemoryDumpStartAddress -= uMemoryDumpStartAddress & uPointerSizeMask; # decrease to align
       uMemoryDumpEndAddress = uMemoryDumpStartAddress + uMemoryDumpSize;
